nlp/requirements.txt
certifi==2026.2.25
charset-normalizer==3.4.7
colorama==0.4.6
filelock==3.28.0
fsspec==2026.3.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
kafka-python==2.3.1
llvmlite==0.47.0
MarkupSafe==3.0.3
more-itertools==11.0.2
mpmath==1.3.0
networkx==3.6.1
numba==0.65.0
numpy==2.4.4
openai-whisper==20250625
pydub==0.25.1
regex==2026.4.4
requests==2.33.1
scikit-learn==1.8.0
scipy==1.17.1
setuptools==81.0.0
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
tiktoken==0.12.0
tqdm==4.67.3
typing_extensions==4.15.0
urllib3==2.6.3
